refac: Rename uitls.py -> utils.py
